pyproject.toml
[project]
name = "smartphones-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "googlesearch-python>=1.3.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-deepseek>=0.1.3",
    "langchain-huggingface>=0.3.0",
    "langchain-neo4j>=0.4.0",
    "mcp[cli]>=1.10.1",
    "neo4j>=5.28.1",
    "pandas>=2.3.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
